requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
redis==3.5.0 # https://github.com/andymccurdy/redis-py
celery==4.4.2  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
django-celery-results~=1.2.1

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/jazzband/django-redis

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5 # https://github.com/Tivix/django-rest-auth/blob/master/docs/index.rst

# My forked additions
# ------------------------------------------------------------------------------
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
# sorl-thumbnail
django-braces==1.14.0
django-ckeditor==5.9.0
django-datatables-too==0.1.3
django-extra-views==0.12.0
django-multi-form-view==2.0.1
django-widget-tweaks==1.4.5

sorl-thumbnail==12.6.3
sorl-thumbnail-serializer-field==0.2.1

stripe==2.48.0